Add files via upload
Traffic congestion is a severe problem in many major cities across the world and it has become a nightmare for the commuters in these cities.
The project is designed to develop a density based dynamic traffic signal system. The signal timing changes automatically on sensing the traffic density at the junction. For this,we used IR sensor and photodiodes pair to detect density of each lane. Depending on the values obtained from respective sensors, system determines which lane is more dense. After determination of more densed lane, system halt it's ongoing operation and schedule delay in a way that densed lane should clears first after it goes back to its regular operation.

Implementation using Arduino Mega Board